move file to parent folder
